Move SSL into TCPClient
