+++ lib/net/tcp_client.rb
@@ -1,7 +1,7 @@
 require 'socket'
+require 'openssl'
 require 'net/tcp_client/version'
 require 'net/tcp_client/exceptions'
-require 'net/tcp_client/ssl_client'
 require 'net/tcp_client/tcp_client'
 # Use the builtin logger only if semantic logger is not already loaded
 require 'net/tcp_client/logging'
